[py visualization] Follow code layout conventions
Bindings should always be added in class-specific helper files, which
are then referred to by the overall module.
