+++ bindings/pydrake/visualization/visualization_py.h
@@ -15,6 +15,9 @@ namespace internal {
 /* Defines bindings per visualization_py_config.cc. */
 void DefineVisualizationConfig(py::module m);
 
+/* Defines bindings per visualization_py_sliders.cc. */
+void DefineVisualizationSliders(py::module m);
+
 }  // namespace internal
 }  // namespace pydrake
 }  // namespace drake
